requirements.txt
numpy==1.24.2
onnx==1.14.1
onnxruntime==1.18.1
onnxruntime_gpu==1.13.1
opencv_python==4.10.0.84
Pillow==11.1.0
